new file:   nbproject/private/config.properties 	modified:   nbproject/private/private.xml 	renamed:    src/org/netgene/ga/core/GenerationResult.java -> src/org/netgene/ga/GenerationResult.java 	renamed:    src/org/netgene/ga/core/GenerationTracker.java -> src/org/netgene/ga/GenerationTracker.java 	renamed:    src/org/netgene/ga/core/GeneticAlgorithm.java -> src/org/netgene/ga/GeneticAlgorithm.java 	renamed:    src/org/netgene/ga/core/GeneticConfiguration.java -> src/org/netgene/ga/GeneticConfiguration.java 	renamed:    src/org/netgene/ga/Individual.java -> src/org/netgene/ga/core/Individual.java 	renamed:    src/org/netgene/ga/Offspring.java -> src/org/netgene/ga/core/Offspring.java 	renamed:    src/org/netgene/ga/Parents.java -> src/org/netgene/ga/core/Parents.java 	renamed:    src/org/netgene/ga/Population.java -> src/org/netgene/ga/core/Population.java 	modified:   src/org/netgene/ga/crossover/CrossoverOperator.java 	modified:   src/org/netgene/ga/crossover/FixedPointCrossover.java 	modified:   src/org/netgene/ga/crossover/HalfPointCrossover.java 	modified:   src/org/netgene/ga/crossover/OnePointCrossover.java 	modified:   src/org/netgene/ga/crossover/Order1Crossover.java 	modified:   src/org/netgene/ga/crossover/TwoPointCrossover.java 	modified:   src/org/netgene/ga/crossover/UniformCrossover.java 	modified:   src/org/netgene/ga/fitness/FitnessFunction.java 	modified:   src/org/netgene/ga/mutator/BitFlipMutator.java 	modified:   src/org/netgene/ga/mutator/GaussianMutator.java 	modified:   src/org/netgene/ga/mutator/IntegerMutator.java 	modified:   src/org/netgene/ga/mutator/InversionMutator.java 	modified:   src/org/netgene/ga/mutator/MutatorOperator.java 	modified:   src/org/netgene/ga/mutator/RandomMutator.java 	modified:   src/org/netgene/ga/mutator/ScrambleMutator.java 	modified:   src/org/netgene/ga/mutator/SwapMutator.java 	modified:   src/org/netgene/ga/selection/parent/CompetitionSelector.java 	modified:   src/org/netgene/ga/selection/parent/ParentSelector.java 	modified:   src/org/netgene/ga/selection/parent/RandomSelector.java 	modified:   src/org/netgene/ga/selection/parent/RankSelector.java 	modified:   src/org/netgene/ga/selection/parent/RouletteSelector.java 	modified:   src/org/netgene/ga/selection/parent/TournamentSelector.java 	modified:   src/org/netgene/ga/stop/StopCondition.java 	modified:   src/org/x/demo/Lesson1.java 	modified:   src/org/x/demo/Lesson2.java 	modified:   src/org/x/demo/Lesson3.java 	modified:   src/org/x/demo/snake/SnakeAI.java 	modified:   src/org/x/demo/snake/SnakeFitness.java 	modified:   src/xsudoku/CustomSwapMutator.java 	modified:   src/xsudoku/SudokuGaSimple.java 	modified:   src/xsudoku/SudokuGaSimple2.java 	modified:   src/xsudoku/SudokuGaSimple3.java 	modified:   src/xsudoku/SudokuGaSimpleTesting.java 	modified:   src/xtest/CustomCrossover.java 	modified:   src/xtest/CustomMutator.java 	modified:   src/xtest/CustomPermMutator.java 	modified:   src/xtest/CustomPyCrossover.java 	modified:   src/xtest/FitnessSharing.java 	modified:   src/xtest/SudokuGA.java 	modified:   src/xtest/SudokuGAPerm.java 	modified:   src/xtest/SudokuGARunner.java 	modified:   src/xtest/SudokuGASimple.java 	modified:   src/xtest/SudokuGA_run.java
