+++ nbproject/private/private.xml
@@ -23,13 +23,9 @@
             <file>file:/C:/storage/priv_cbaba_dev/src/xtest/SudokuGARunner.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/mutator/IntegerMutator.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/mutator/ScrambleMutator.java</file>
-            <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/Parents.java</file>
-            <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/core/GeneticAlgorithm.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/xtest/Sudoku.java</file>
-            <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/core/GenerationTracker.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/crossover/CrossoverOperator.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/utils/Printers.java</file>
-            <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/Population.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/selection/parent/TournamentSelector.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/chromosome/NumericChromosome.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/utils/Printer.java</file>
@@ -39,6 +35,7 @@
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/utils/RandomUtils.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/x/demo/snake/TestSnake.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/chromosome/PermutationChromosome.java</file>
+            <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/GeneticAlgorithm.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/stop/StopCondition.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/xsudoku/Sudoku.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/utils/TaskExecutor.java</file>
@@ -47,27 +44,26 @@
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/chromosome/Chromosome.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/selection/parent/CompetitionSelector.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/xsudoku/SudokuGaSimple3.java</file>
-            <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/core/GeneticConfiguration.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/x/demo/Lesson2.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/chromosome/DoubleChromosome.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/xtest/FitnessSharing.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/exception/GaException.java</file>
+            <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/GeneticConfiguration.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/xtest/SudokuGASimple.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/crossover/TwoPointCrossover.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/mutator/BitFlipMutator.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/x/demo/snake/PrintWorld.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/exception/CrossoverException.java</file>
+            <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/GenerationTracker.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/gene/NumericGene.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/exception/MutatorException.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/xsudoku/SudokuGaSimpleTesting.java</file>
-            <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/Individual.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/gene/DoubleGene.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/x/demo/snake/SnakeAI.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/chromosome/IntegerChromosome.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/x/demo/Lesson3.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/xsudoku/CustomSwapMutator.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/network/MultiLayerNetwork.java</file>
-            <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/Offspring.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/crossover/UniformCrossover.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/x/demo/snake/SnakeFitness.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/crossover/Order1Crossover.java</file>
@@ -79,14 +75,15 @@
             <file>file:/C:/storage/priv_cbaba_dev/src/xtest/SudokuGA_run.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/xtest/CustomPyCrossover.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/network/learning/optimizer/Adam.java</file>
+            <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/core/Individual.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/x/demo/snake/Node.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/crossover/OnePointCrossover.java</file>
-            <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/core/GenerationResult.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/fitness/FitnessFunction.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/exception/SelectionException.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/x/demo/Lesson1.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/mutator/MutatorOperator.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/mutator/RandomMutator.java</file>
+            <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/GenerationResult.java</file>
             <file>file:/C:/storage/priv_cbaba_dev/src/org/netgene/ga/mutator/InversionMutator.java</file>
         </group>
     </open-files>
