+++ src/org/netgene/ga/crossover/OnePointCrossover.java
@@ -21,9 +21,9 @@
 package org.netgene.ga.crossover;
 
 import java.io.Serializable;
-import org.netgene.ga.Individual;
+import org.netgene.ga.core.Individual;
 import org.netgene.ga.chromosome.Chromosome;
-import org.netgene.ga.Offspring;
+import org.netgene.ga.core.Offspring;
 import org.netgene.ga.chromosome.PermutationChromosome;
 import org.netgene.ga.exception.CrossoverException;
 import org.netgene.utils.RandomUtils;
