+++ src/org/netgene/ga/mutator/RandomMutator.java
@@ -6,7 +6,7 @@
 package org.netgene.ga.mutator;
 
 import java.io.Serializable;
-import org.netgene.ga.Individual;
+import org.netgene.ga.core.Individual;
 import org.netgene.ga.chromosome.DoubleChromosome;
 import org.netgene.ga.chromosome.IntegerChromosome;
 import org.netgene.ga.chromosome.NumericChromosome;
