+++ src/org/netgene/ga/selection/parent/CompetitionSelector.java
@@ -6,8 +6,8 @@
 package org.netgene.ga.selection.parent;
 
 import java.io.Serializable;
-import org.netgene.ga.Individual;
-import org.netgene.ga.Population;
+import org.netgene.ga.core.Individual;
+import org.netgene.ga.core.Population;
 import org.netgene.ga.exception.SelectionException;
 //
