+++ src/org/netgene/ga/selection/parent/ParentSelector.java
@@ -21,9 +21,9 @@
 package org.netgene.ga.selection.parent;
 
 import java.io.Serializable;
-import org.netgene.ga.Individual;
-import org.netgene.ga.Population;
-import org.netgene.ga.Parents;
+import org.netgene.ga.core.Individual;
+import org.netgene.ga.core.Population;
+import org.netgene.ga.core.Parents;
 import org.netgene.ga.exception.SelectionException;
 
 /**
