+++ src/org/netgene/ga/selection/parent/TournamentSelector.java
@@ -22,8 +22,8 @@ package org.netgene.ga.selection.parent;
 
 import java.io.Serializable;
 import java.util.stream.Stream;
-import org.netgene.ga.Individual;
-import org.netgene.ga.Population;
+import org.netgene.ga.core.Individual;
+import org.netgene.ga.core.Population;
 import org.netgene.ga.exception.SelectionException;
 
 /**
