+++ src/xsudoku/CustomSwapMutator.java
@@ -6,7 +6,7 @@
 package xsudoku;
 
 import java.util.ArrayList;
-import org.netgene.ga.Individual;
+import org.netgene.ga.core.Individual;
 import org.netgene.ga.chromosome.IntegerChromosome;
 import org.netgene.ga.exception.MutatorException;
 import org.netgene.ga.gene.IntegerGene;
