+++ src/xsudoku/SudokuGaSimple3.java
@@ -8,12 +8,12 @@ package xsudoku;
 import java.util.ArrayList;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import org.netgene.ga.Individual;
-import org.netgene.ga.Population;
+import org.netgene.ga.core.Individual;
+import org.netgene.ga.core.Population;
 import org.netgene.ga.chromosome.IntegerChromosome;
-import org.netgene.ga.core.GenerationTracker;
-import org.netgene.ga.core.GeneticAlgorithm;
-import org.netgene.ga.core.GeneticConfiguration;
+import org.netgene.ga.GenerationTracker;
+import org.netgene.ga.GeneticAlgorithm;
+import org.netgene.ga.GeneticConfiguration;
 import org.netgene.ga.crossover.*;
 import org.netgene.ga.fitness.FitnessFunction;
 import org.netgene.ga.gene.IntegerGene;
