+++ src/xtest/CustomPermMutator.java
@@ -6,7 +6,7 @@
 package xtest;
 
 import java.util.ArrayList;
-import org.netgene.ga.Individual;
+import org.netgene.ga.core.Individual;
 import org.netgene.ga.chromosome.Chromosome;
 import org.netgene.ga.chromosome.IntegerChromosome;
 import org.netgene.ga.exception.MutatorException;
