+++ src/xtest/FitnessSharing.java
@@ -5,7 +5,7 @@
  */
 package xtest;
 
-import org.netgene.ga.Population;
+import org.netgene.ga.core.Population;
 import org.netgene.ga.chromosome.IntegerChromosome;
 
 /**
